fix: SignUpDoneForm 제거 및 page.tsx 구조에 통합
